requirements.txt
selenium
google-generativeai
ipython
flask
pillow
python-dotenv